[BREAKING][core] move single_controller into verl directory
