+++ tests/ray/detached_worker/server.py
@@ -25,10 +25,10 @@ import torch
 from torch import nn
 
 import ray
-from single_controller.ray import RayClassWithInitArgs, RayResourcePool
-from single_controller.ray.megatron import NVMegatronRayWorkerGroup
-from single_controller.base.megatron.worker import MegatronWorker
-from single_controller.ray.decorator import register, Dispatch
+from verl.single_controller.ray import RayClassWithInitArgs, RayResourcePool
+from verl.single_controller.ray.megatron import NVMegatronRayWorkerGroup
+from verl.single_controller.base.megatron.worker import MegatronWorker
+from verl.single_controller.ray.decorator import register, Dispatch
 from verl import DataProto
 from verl.models.llama.megatron import ParallelLlamaForCausalLMRmPadPP
