+++ verl/single_controller/base/worker_group.py
@@ -20,7 +20,7 @@ import signal
 import time
 from typing import List, Any, Callable, Dict
 
-from single_controller.base.decorator import MAGIC_ATTR, Dispatch, get_predefined_dispatch_fn, get_predefined_execute_fn
+from verl.single_controller.base.decorator import MAGIC_ATTR, Dispatch, get_predefined_dispatch_fn, get_predefined_execute_fn
 
 
 class ResourcePool:
